fix: fix bugs preventing postman tests from passing
Fix mistakes in sql queries. Film class must have Genres and Mpa as objects, not Integer

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -24,7 +24,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenreById(@PathVariable Long id) {
+    public Genre getGenreById(@PathVariable Integer id) {
         log.info("received HTTP request to get genre by id {}", id);
         return genreService.findGenreOrThrow(id);
     }
